Fixes to comparison

+++ src/main/java/com/kcthota/version4j/comparison/VersionComparator.java
@@ -11,9 +11,9 @@ import com.kcthota.version4j.models.Version;
 public class VersionComparator implements Comparator<Version> {
 	
 	public int compare(Version v1, Version v2) {
-		if (v1.getScalarValue() > v2.getScalarValue()) {
+		if (v1.getScalarValue().doubleValue() > v2.getScalarValue().doubleValue()) {
 			return 1;
-		} else if (v1.getScalarValue() == v2.getScalarValue()) {
+		} else if (v1.getScalarValue().doubleValue() == v2.getScalarValue().doubleValue()) {
 			return 0;
 		} else {
 			return -1;
